Fixed byte/int bug
Former-commit-id: de189dd212687668dba5b52315fe8285a1871cb8

+++ common-tools/clas-reco/src/main/java/org/jlab/clas/detector/DetectorData.java
@@ -420,7 +420,7 @@ public class DetectorData {
            int nrows = bank.rows();
            
            for(int row = 0; row < nrows; row++){
-               int charge = bank.getByte("q", row);
+               int charge = bank.getInt("q", row);
                
                double p = bank.getFloat("p",row);
                //System.out.println("Central Detector MOmentum  " + p);
